pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.0.1",
    "screeninfo>=0.8.1",
    "ultralytics>=8.3.23",
    "websockets>=15.0",
]
